pyproject.toml
[tool.poetry]
name = "Ask me anything about the uploaded PDF document"
version = "0.1.0"
description = "Dialogue with your Documents for Data-Driven Decision Making"
authors = ["Josoroma <pablo@josoroma.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.25.0"
streamlit-chat = "^0.1.1"
streamlit-extras = "^0.3.0"
langchain = "^0.0.250"
openai = "^0.27.8"
tiktoken = "^0.4.0"
pypdf = "^3.14.0"
pinecone-client = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
